+++ smartcontract/sdk/rs/src/commands/tenant/create.rs
@@ -13,6 +13,8 @@ pub struct CreateTenantCommand {
     pub code: String,
     pub administrator: Pubkey,
     pub token_account: Option<Pubkey>,
+    pub metro_route: bool,
+    pub route_aliveness: bool,
 }
 
 impl CreateTenantCommand {
@@ -33,6 +35,8 @@ impl CreateTenantCommand {
                     code,
                     administrator: self.administrator,
                     token_account: self.token_account,
+                    metro_route: self.metro_route,
+                    route_aliveness: self.route_aliveness,
                 }),
                 vec![
                     AccountMeta::new(pda_pubkey, false),
@@ -66,6 +70,8 @@ mod tests {
                     code: "test".to_string(),
                     administrator,
                     token_account: None,
+                    metro_route: true,
+                    route_aliveness: false,
                 })),
                 predicate::always(),
             )
@@ -75,6 +81,8 @@ mod tests {
             code: "test/invalid".to_string(),
             administrator: Pubkey::default(),
             token_account: None,
+            metro_route: true,
+            route_aliveness: false,
         }
         .execute(&client);
 
@@ -84,6 +92,8 @@ mod tests {
             code: "test".to_string(),
             administrator,
             token_account: None,
+            metro_route: true,
+            route_aliveness: false,
         }
         .execute(&client);